% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_error.R
\name{bootstrap_error}
\alias{bootstrap_error}
\title{Bootstrap the standard deviation of the mean}
\usage{
bootstrap_error(x, N = 999, printit = F)
}
\arguments{
\item{x}{A vector of deviations from the mean (or any numbers).}

\item{N}{An integer. The number of samplings to perform.}

\item{printit}{Logical. Whether to print the iteration to the prompt.}
}
\value{
The standard deviation of the re-calculated means of all bootstrapped samples.
}
\description{
This function performs a bootstrap on a vector of deviations from means along a given variable to measure the error of the mean
}
\author{
Raphael Scherrer
}
