% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pulsr_results.R
\name{combine_pulsr_results}
\alias{combine_pulsr_results}
\title{Combine output files}
\usage{
combine_pulsr_results(path)
}
\arguments{
\item{path}{Path to where the output files are. This is the parent directory to the folders corresponding to each dataset: "males_wholebird", "females_wholebird", "males_perpatch", "females_perpatch". They must all be present. Each contains as many folders as there are variables. Each of these variable-folders contains an RDS file, which was saved by the function \code{fit_pulsr} on the cluster and contains the results of the fitting.}
}
\description{
This function combines output files from pulsR analyses from different variables. There is no argument. For each dataset, it combines the output lists generated for each variable into a single RDS file.
}
\note{
This functions saves an RDS file into each dataset-folder, named "results_all_variables.rds".
}
\author{
Raphael Scherrer
}
