% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_input}
\alias{create_input}
\title{Create PEPHYLO input}
\usage{
create_input(inpath, outpath, wholeBird = F, varNames = c("VS.v", "S.v",
  "M.v", "L.v"), whatSex = "M", saveit = T, plotit = F)
}
\arguments{
\item{inpath}{A string. The path to the folder where the input file "spec.csv" is to be looked for.}

\item{outpath}{A string. The path to the folder where the output table is to be written.}

\item{wholeBird}{Logical. Whether the points in color space represent whole birds (\code{TRUE}) or patches (\code{FASLE}).}

\item{varNames}{Vector of strings. The name(s) of the variable(s) from the reflectance data to include in the Principal Component Analysis}

\item{whatSex}{A character. \code{"M"} for males, \code{"F"} for females.}

\item{saveit}{Logical. Whether to save the output table in a file.}

\item{plotit}{Logical. Whether to plot PCA eigenvalues.}
}
\value{
A data frame with each sex of each species in rows, and principal components in columns. If \code{wholeBird = TRUE}, columns represent loadings of each patch on principal components 2, 3 and 4. Otherwise, columns represent principal components 2, 3 and 4.
}
\description{
This function creates a table of traits per species suitable for input in Bokma's PEPHYLO
}
\details{
If \code{wholeBird = TRUE}, the different reflectance measurements for each patch are inputs of the PCA. The PC loadings describe the average position of each species in this highly multidimensional phenotypic space. If \code{wholeBird = FALSE}, the reflectance measurements are the inputs of the PCA. The PC loadings represent the position of each patch into what is closer to a true color space (a few dimensions). But PEPHYLO requires observed phenotypes on a per-species basis, so the PC loadings of each patch are converted into dependent variables.
}
\references{
Bokma, F. (2008). Detection of “punctuated equilibrium” by Bayesian estimation of speciation and extinction rates, ancestral character states, and rates of anagenetic and cladogenetic evolution on a molecular phylogeny. Evolution, 62(11), 2718–2726. https://doi.org/10.1111/j.1558-5646.2008.00492.x
}
\author{
Raphael Scherrer
}
