% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dematlabilize_data.R
\name{dematlabilize_data}
\alias{dematlabilize_data}
\title{Export Matlab output}
\usage{
dematlabilize_data(path = ".")
}
\arguments{
\item{path}{Path to where the output files are.}
}
\value{
A huge data frame with MCMC variables in columns and iterations in rows.
}
\description{
This function turns matlab output into a data frame handily usable in R
}
\author{
Raphael Scherrer
}
