% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error.R
\name{estimate_error}
\alias{estimate_error}
\title{Estimate error on the means}
\usage{
estimate_error(speciesData, indivData, nPC)
}
\arguments{
\item{speciesData}{A data frame with PC loadings and metadata at the species level.}

\item{indivData}{A data frame with PC loadings and metadata at the individual level.}

\item{nPC}{How many PC to retain? Either an integer i, then PC 1 to i will be retained, or a vector of integers representing what PCs to retain.}
}
\value{
A vector of estimated errors, one for each dependent variable defined by \code{nPC}.
}
\description{
This function estimates errors made on measuring the mean by calculating the standard deviation of a bootstrapped distribution.
}
\author{
Raphael Scherrer
}
