% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_spext.R
\name{estimate_spext}
\alias{estimate_spext}
\title{Esimate speciation and extinction}
\usage{
estimate_spext(treepath)
}
\arguments{
\item{treepath}{A path to where the tree file is. The tree file should be named "Tress.txt" and be in Newick format. It may contain several trees.}
}
\value{
A vector with a speciation and an extinction rate.
}
\description{
This function fits Nee's likelihood function to one or more trees and returns the estimated speciation and extinction rates by maximum likelihood. If several trees are input, the average of those is returned.
}
\note{
\code{ape} is a dependency.
}
\author{
Raphael Scherrer
}
