% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pulsr.R
\name{fit_pulsr}
\alias{fit_pulsr}
\title{Function to test all 9 models on a phenotypic vector, nrepet times}
\usage{
fit_pulsr(trait_vector, tree, species_names, model_types, nrepet)
}
\arguments{
\item{trait_vector}{A vector of phenotypic values across species.}

\item{tree}{A phylogeny, in ape format.}

\item{species_names}{A vector of species names.}

\item{model_types}{The names of the models to fit, in a vector. See the scripts downloaded from pulsr's github (stored in analysis2/R) for a list of models available.}

\item{nrepet}{The number of replicate analyses.}
}
\value{
A list of lists of outputs of the fitting function. Levels of nestedness: model type, replicates.
}
\description{
This function fits (potentially) several models of evolution on a phenotypic vector using Levy processes from the package pulsr. Each fitting can be replicated to ensure convergence to a single optimum likelihood.
}
\note{
This function was run on the cluster. It is made to call the functions in pulsr, which are not yet a working package but a bunch of scripts in a folder associated. So, this function won't work unless these scripts are availeble (see the source of the function to know where they should be). Also, this function runs on a single variable and not on a dataset of multiple variables because some variable cause convergence problems. This avoids the whole analysis to break down because of a single problematic variable in a multivariate dataset.
}
\author{
Raphael Scherrer
}
