% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_matching_names.R
\name{is_matching_names}
\alias{is_matching_names}
\title{View PEPHYLO input files}
\usage{
is_matching_names(outpath)
}
\arguments{
\item{outpath}{A string. The path to the folder where the output file is to be saved.}
}
\value{
Messages summarizing what species are missing, if any.
}
\description{
This function checks the correspondance between species names in the data and in the phylogeny(ies)
}
\author{
Raphael Scherrer
}
