% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_pephylo.R
\name{make_prior_pephylo}
\alias{make_prior_pephylo}
\title{Anagenetic and cladogenetic rate priors for PEPHYLO}
\usage{
make_prior_pephylo(meantraits, age, outpath)
}
\arguments{
\item{meantraits}{A matrix or data frame with mean traits for each species (rows), for each dependent variable (columns).}

\item{age}{The age of the clade (in Myr).}

\item{outpath}{A string. The path to the folder where the output file is to be saved.}
}
\value{
Returns 1 if succeeds. Saves the output file into the specified folder.
}
\description{
This function writes a table with prior rates of anagenetic and cladogenetic evolution. The prior for sa2 is calculated by dividing interspecific variance for each dependent variable by the age of the clade. The prior for sc2 is calculated by dividing the same interspecfic variance by the number of species. Saves an output file with four colums filled with zeros (unwanted parameters), plus two columns with sa2 and sc2 rate priors. One row per dependent variable.
}
\note{
The amount of evolutionary change along a branch in a phylogeny is assumed to be normally distributed with mean zero and variance sa2 t + sc2 S, with t the age of the clade and S the number of speciation events. sa2 represents the rate of change expected over 1 Myr. sc2 represents the rate of change upon speciation e.g. change with mean zero and variance sc2 upon speciation.
}
\author{
Raphael Scherrer
}
