% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spname_pephylo.R
\name{make_spname_pephylo}
\alias{make_spname_pephylo}
\title{Write species names for PEPHYLO}
\usage{
make_spname_pephylo(speciesNames, outpath)
}
\arguments{
\item{speciesNames}{A vector of species names.}

\item{outpath}{A string. The path to the folder where the output file is to be saved.}
}
\value{
Returns 1 if succeeds. Saves the output file into the specified folder.
}
\description{
This function writes a vector of species names into a file for input in PEPHYLO
}
\author{
Raphael Scherrer
}
