% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_std_pephylo.R
\name{make_std_pephylo}
\alias{make_std_pephylo}
\title{Write PEPHYLO standard deviation input}
\usage{
make_std_pephylo(errors, nrows, outpath, whatSex)
}
\arguments{
\item{errors}{A vector of standard deviations (errors on the means) for each dependent variable. The errors can be calculated with \code{bootrstrap_error}.}

\item{nrows}{An integer. The number of rows of the input table.}

\item{outpath}{A string. The path to the folder where the output file is to be saved.}

\item{whatSex}{A logical vector. The indices of the observations that are of the sex of interest.}
}
\value{
Returns 1 if succeeds. Saves the output file into the specified folder.
}
\description{
This function saves the standard deviations of the means required for running PEPHYLO
}
\author{
Raphael Scherrer
}
