% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_perpatch.R
\name{pca_perpatch}
\alias{pca_perpatch}
\title{Project bird species in patch PC space}
\usage{
pca_perpatch(inpath, varNames = c("VS.v", "S.v", "M.v", "L.v"))
}
\arguments{
\item{inpath}{A string. The path to the folder where to find the input data file. Note: the input data file is named "specs.csv".}

\item{varNames}{A vector of strings. The names of the dependent variables in the input data frame.}
}
\value{
A list of two elements. The first is the output of \code{prcomp} performed on the dependent variables, where each bird has coordinates. The second element a table with metadata identifying each observation in the PCA.
}
\description{
This function performs a Principal Component Analysis from quantum catches measured on different patches on the plumage. Each bird is represented by several patches, so several points.
}
\author{
Raphael Scherrer
}
