% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_by_patch.R
\name{rearrange_by_patch}
\alias{rearrange_by_patch}
\title{Rearrange with patches in columns}
\usage{
rearrange_by_patch(df)
}
\arguments{
\item{df}{A data frame. Must have a column "patch" and a column "species".}
}
\value{
A rearranged data frame with different patches placed in different columns.
}
\description{
This function rearranges a dataframe with patches in different columns
}
\author{
Raphael Scherrer
}
