% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange_indiv_by_patch.R
\name{rearrange_indiv_by_patch}
\alias{rearrange_indiv_by_patch}
\title{Rearrange individuals with patches in columns}
\usage{
rearrange_indiv_by_patch(indivData)
}
\arguments{
\item{indivData}{A data frame. Must have columns "patch", "species", "sex" and "specimen".}
}
\value{
A rearranged data frame with different patches placed in different columns.
}
\description{
This function rearranges an individual-level dataset by patch
}
\note{
It is a separate function from \code{rearrange_by_patch} because it uses a for-loop and takes longer. For rearranging a species-level data frame, use \code{rearrange_by_patch}.
}
\author{
Raphael Scherrer
}
