% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad_patches.R
\name{remove_bad_patches}
\alias{remove_bad_patches}
\title{Remove patches with NAs}
\usage{
remove_bad_patches(specTab)
}
\arguments{
\item{specTab}{A data frame. Must have a column "patch" and a column "species".}
}
\value{
A subset of the input data frame with only patches that have no missing data.
}
\description{
This function removes the patches that are not measured in all species or all sexes from a data frame
}
\author{
Raphael Scherrer
}
