% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_error.R
\name{estimate_error}
\alias{estimate_error}
\title{Estimate error on the means}
\usage{
estimate_error(speciesPC, indivPC, metaSpecies, metaIndiv, nPC)
}
\arguments{
\item{speciesPC}{A matrix containing the PC loadings for each species.}

\item{indivPC}{A matrix containing the PC loadings for each individual. Should have the same number of columns as \code{speciesPC}.}

\item{metaSpecies}{A data frame containing species-level metadata, used to assign individuals to species.}

\item{metaIndiv}{A data frame containing individual-level metadata.}

\item{nPC}{How many PC to retain? Either an integer i, then PC 1 to i will be retained, or a vector of integers representing what PCs to retain.}
}
\value{
A vector of estimated errors, one for each dependent variable defined by \code{nPC}.
}
\description{
This function estimates errors made on measuring the mean by calculating the standard deviation of a bootstrapped distribution.
}
\author{
Raphael Scherrer
}
