% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_individuals.R
\name{project_individuals}
\alias{project_individuals}
\title{Project individuals into PC space}
\usage{
project_individuals(inpath, pca)
}
\arguments{
\item{inpath}{A string. The path to the folder where to find the input data file. Note: the input data file is named "specs_indiv.csv".}

\item{pca}{A list with fields \code{input} and \code{output}, where \code{output} is a \code{prcomp} object. Typically the output of \code{pca_wholebird} or \code{pca_perpatch}.}
}
\value{
A matrix of principal component coordinates.
}
\description{
This projects individual specimens into the PC space defined by species-averages.
}
\author{
Raphael Scherrer
}
