% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_individuals.R
\name{project_individuals}
\alias{project_individuals}
\title{Project individuals into PC space}
\usage{
project_individuals(inpath, pca, varNames = c("VS.v", "S.v", "M.v", "L.v"))
}
\arguments{
\item{inpath}{A string. The path to the folder where to find the input data file. Note: the input data file is named "specs_indiv.csv".}

\item{pca}{A \code{prcomp} object.}

\item{varNames}{A vector of strings. The names of the dependent variables in the input data frame.}
}
\value{
A data frame of principal component coordinates and metadata.
}
\description{
This projects individual specimens into the PC space defined by species-averages.
}
\author{
Raphael Scherrer
}
